introduced attached property. only fire mount/unmount and send native_patches if attached
